final solution(2.1)

+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserServiceImpl.java
@@ -67,6 +67,9 @@ public class UserServiceImpl implements UserService, UserDetailsService {
     @Transactional
     @Override
     public void deleteById(Long id) {
+        if (userRepository.findUserById(id) == null) {
+            throw new UserNotFoundException();
+        }
         userRepository.deleteById(id);
     }
